users/requirements.txt
attrs==21.2.0
grpcio==1.39.0
grpcio-tools==1.39.0
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
protobuf==3.17.3
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
six==1.16.0
toml==0.10.2
